treewide: Change DOUBLEs used for aggregation into DECIMAL(38, 2).
Sums of DOUBLE are imprecise and cannot be implemented linearly without
extending precision.  Although that is possible, e.g. see
https://github.com/feldera/feldera/issues/4360, it is even easier to use
DECIMAL instead, in many cases.  This commit changes all of the instances
of DOUBLE that are potentially used for aggregation into DECIMAL.

Some DOUBLE values cannot sensibly be used for arithmetical aggregation.
This commit does not change them.  (For example, it doesn't make sense
to take the sum, average, or standard deviation over latitude and
longitude values.)

+++ demo/packaged/sql/01-fraud-detection.sql
@@ -70,7 +70,7 @@ CREATE TABLE CUSTOMER (
 -- Credit card transactions.
 CREATE TABLE TRANSACTION (
     ts TIMESTAMP LATENESS INTERVAL 10 MINUTES, -- Transaction time
-    amt DOUBLE,                                -- Transaction amount
+    amt DECIMAL(38, 2),                        -- Transaction amount
     cc_num BIGINT NOT NULL,                    -- Credit card number
     shipping_lat DOUBLE,                       -- Shipping address latitude
     shipping_long DOUBLE,                      -- Shipping address longitude
